refactor(backup): consolidate into features/backup domain module
