--- src/components/backup/index.ts
@@ -1,3 +1,0 @@
-export { default as BackupCloudStep } from './BackupCloudStep';
-export { default as BackupSheetSection } from './BackupSheetSection';
-export { default as RestoreCloudStep } from './RestoreCloudStep';
