+++ src/features/backup/backup.ts
@@ -7,6 +7,7 @@ import { endsWith } from 'lodash';
 import { analytics } from '@/analytics';
 import { Alert as NativeAlert } from '@/components/alerts';
 import { IS_ANDROID, IS_DEV } from '@/env';
+import AesEncryptor from '@/handlers/aesEncryption';
 import { authenticateWithPIN, decryptPIN, maybeAuthenticateWithPINAndCreateIfNeeded } from '@/handlers/authentication';
 import {
   CLOUD_BACKUP_ERRORS,
@@ -20,24 +21,23 @@ import {
 } from '@/handlers/cloudBackup';
 import { WrappedAlert as Alert } from '@/helpers/alert';
 import walletBackupStepTypes from '@/helpers/walletBackupStepTypes';
+import WalletBackupTypes from '@/helpers/walletBackupTypes';
 import { getUserError } from '@/hooks/useWalletCloudBackup';
 import * as kc from '@/keychain';
 import * as i18n from '@/languages';
 import { logger, RainbowError } from '@/logger';
 import * as keychain from '@/model/keychain';
+import { clearAllStorages } from '@/model/mmkv';
+import { getRemoteConfig } from '@/model/remoteConfig';
+import { createWallet, type AllRainbowWallets, type RainbowWallet } from '@/model/wallet';
 import Navigation from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
-import { backupsStore, CloudBackupState } from '@/state/backups/backups';
 import { loadWallets, refreshWalletInfo, setAllWalletsWithIdsAsBackedUp } from '@/state/wallets/walletsStore';
 import { identifierForVendorKey, pinKey, privateKeyKey, seedPhraseKey } from '@/utils/keychainConstants';
 import { openInBrowser } from '@/utils/openInBrowser';
 import { cloudPlatform } from '@/utils/platform';
 
-import AesEncryptor from '../handlers/aesEncryption';
-import WalletBackupTypes from '../helpers/walletBackupTypes';
-import { clearAllStorages } from './mmkv';
-import { getRemoteConfig } from './remoteConfig';
-import { createWallet, type AllRainbowWallets, type RainbowWallet } from './wallet';
+import { backupsStore, CloudBackupState } from './stores/backupsStore';
 
 const { DeviceUUID } = NativeModules;
 const encryptor = new AesEncryptor();
