+++ src/features/backup/components/BackupSheetSection.tsx
@@ -1,16 +1,15 @@
 import React, { Fragment, useEffect } from 'react';
 
 import { analytics } from '@/analytics';
+import { RainbowButton } from '@/components/buttons';
+import { Column, ColumnWithMargins } from '@/components/layout';
+import { SheetActionButton } from '@/components/sheet';
+import { Text } from '@/components/text';
 import { Bleed, Separator } from '@/design-system';
 import styled from '@/framework/ui/styled-thing';
 import { opacity } from '@/framework/ui/utils/opacity';
 import { padding } from '@/styles';
-
-import { useTheme } from '../../theme/ThemeContext';
-import { RainbowButton } from '../buttons';
-import { Column, ColumnWithMargins } from '../layout';
-import { SheetActionButton } from '../sheet';
-import { Text } from '../text';
+import { useTheme } from '@/theme/ThemeContext';
 
 const Footer = styled(ColumnWithMargins).attrs({
   margin: 19,
