+++ src/features/backup/components/CloudBackupPrompt.tsx
@@ -1,45 +1,49 @@
 import React, { useCallback } from 'react';
 
+import { format } from 'date-fns';
 import { type Source } from 'react-native-fast-image';
 
-import ManuallyBackedUpIcon from '@/assets/ManuallyBackedUp.png';
+import WalletsAndBackupIcon from '@/assets/WalletsAndBackup.png';
+import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
+import { ImgixImage } from '@/components/images';
 import { Bleed, Box, Inline, Inset, Separator, Stack, Text } from '@/design-system';
-import walletBackupTypes from '@/helpers/walletBackupTypes';
-import walletTypes from '@/helpers/walletTypes';
 import * as i18n from '@/languages';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
 import { useSelectedWallet } from '@/state/wallets/walletsStore';
+import { cloudPlatform } from '@/utils/platform';
 
-import ButtonPressAnimation from '../animations/ButtonPressAnimation';
-import { ImgixImage } from '../images';
+import { executeFnIfCloudBackupAvailable } from '../backup';
+import { useCreateBackup } from '../hooks/useCreateBackup';
+import { backupsStore } from '../stores/backupsStore';
 
 const imageSize = 72;
 
-export default function ManualBackupPrompt() {
+export default function CloudBackupPrompt() {
   const { navigate, goBack } = useNavigation();
+  const mostRecentBackup = backupsStore(state => state.mostRecentBackup);
   const selectedWallet = useSelectedWallet();
+  const createBackup = useCreateBackup();
 
-  const onManualBackup = async () => {
-    if (!selectedWallet) {
-      return;
-    }
-    const title =
-      selectedWallet?.imported && selectedWallet.type === walletTypes.privateKey
-        ? (selectedWallet.addresses || [])[0].label
-        : selectedWallet.name;
+  const onCloudBackup = useCallback(() => {
+    if (!selectedWallet) return;
 
+    // pop the bottom sheet, and navigate to the backup section inside settings sheet
     goBack();
     navigate(Routes.SETTINGS_SHEET, {
-      screen: Routes.SECRET_WARNING,
-      params: {
-        isBackingUp: true,
-        title,
-        backupType: walletBackupTypes.manual,
-        walletId: selectedWallet.id,
-      },
+      screen: Routes.SETTINGS_SECTION_BACKUP,
+      initial: false,
     });
-  };
+
+    executeFnIfCloudBackupAvailable({
+      fn: () =>
+        createBackup({
+          walletId: selectedWallet.id,
+          addToCurrentBackup: true,
+        }),
+      logout: true,
+    });
+  }, [createBackup, goBack, navigate, selectedWallet]);
 
   const onMaybeLater = useCallback(() => goBack(), [goBack]);
 
@@ -55,12 +59,12 @@ export default function ManualBackupPrompt() {
             marginRight={{ custom: -12 }}
             marginTop={{ custom: 0 }}
             marginBottom={{ custom: 8 }}
-            source={ManuallyBackedUpIcon as Source}
+            source={WalletsAndBackupIcon as Source}
             width={{ custom: imageSize }}
             size={imageSize}
           />
           <Text align="center" size="26pt" weight="bold" color="label">
-            {i18n.t(i18n.l.back_up.manual.backup_manually_now)}
+            {i18n.t(i18n.l.back_up.cloud.add_wallet_to_cloud_backups)}
           </Text>
         </Stack>
       </Inset>
@@ -69,12 +73,15 @@ export default function ManualBackupPrompt() {
         <Separator color="separatorSecondary" thickness={1} />
       </Bleed>
 
-      <ButtonPressAnimation scaleTo={0.95} onPress={onManualBackup}>
+      <ButtonPressAnimation scaleTo={0.95} onPress={onCloudBackup}>
         <Box alignItems="center" justifyContent="center" paddingTop={'24px'} paddingBottom={'24px'}>
           <Box alignItems="center" justifyContent="center" width="full">
             <Inline alignHorizontal="justify" alignVertical="center" wrap={false}>
               <Text color={'action (Deprecated)'} size="20pt" weight="bold">
-                {i18n.t(i18n.l.back_up.manual.back_up_now)}
+                􀎽{' '}
+                {i18n.t(i18n.l.back_up.cloud.back_to_cloud_platform_now, {
+                  cloudPlatform,
+                })}
               </Text>
             </Inline>
           </Box>
@@ -90,7 +97,7 @@ export default function ManualBackupPrompt() {
           <Box alignItems="center" justifyContent="center" width="full">
             <Inline alignHorizontal="justify" alignVertical="center" wrap={false}>
               <Text color={'labelSecondary'} size="20pt" weight="bold">
-                {i18n.t(i18n.l.back_up.manual.already_backed_up)}
+                {i18n.t(i18n.l.back_up.cloud.mayber_later)}
               </Text>
             </Inline>
           </Box>
@@ -100,6 +107,20 @@ export default function ManualBackupPrompt() {
       <Bleed horizontal="24px">
         <Separator color="separatorSecondary" thickness={1} />
       </Bleed>
+
+      {mostRecentBackup && (
+        <Box alignItems="center" justifyContent="center" paddingTop={'24px'} paddingBottom={'24px'}>
+          <Box alignItems="center" justifyContent="center" width="full">
+            <Inline alignHorizontal="justify" alignVertical="center" wrap={false}>
+              <Text color={'labelTertiary'} size="15pt" weight="medium">
+                {i18n.t(i18n.l.back_up.cloud.latest_backup, {
+                  date: format(new Date(mostRecentBackup.lastModified), "M/d/yy 'at' h:mm a"),
+                })}
+              </Text>
+            </Inline>
+          </Box>
+        </Box>
+      )}
     </Inset>
   );
 }
