+++ src/hooks/useImportingWallet.ts
@@ -7,6 +7,7 @@ import { analytics } from '@/analytics';
 import useExperimentalFlag, { PROFILES } from '@/config/experimentalHooks';
 import { IS_ANDROID, IS_TEST } from '@/env';
 import { isValidAddress, looksLikeAddress } from '@/features/address/core/address';
+import { backupsStore } from '@/features/backup/stores/backupsStore';
 import { fetchENSAvatar } from '@/features/ens/hooks/useENSAvatar';
 import { fetchReverseRecord } from '@/features/ens/utils/handlers';
 import { getProvider, isValidBluetoothDeviceId, resolveUnstoppableDomain } from '@/handlers/web3';
@@ -22,7 +23,6 @@ import { navigateAfterOnboarding } from '@/navigation/onboardingNavigation';
 import Routes from '@/navigation/routesNames';
 import { type ImportFlowContext } from '@/navigation/types';
 import { ChainId } from '@/state/backendNetworks/types';
-import { backupsStore } from '@/state/backups/backups';
 import { walletLoadingStore } from '@/state/walletLoading/walletLoading';
 import { useAccountAddress, useWallets } from '@/state/wallets/walletsStore';
 import { sanitizeSeedPhrase } from '@/utils/formatters';
