+++ src/navigation/types.ts
@@ -18,6 +18,7 @@ import { type ParsedAddressAsset } from '@/entities/tokens';
 import { type PendingTransaction, type RainbowTransaction } from '@/entities/transactions';
 import { type UniqueAsset } from '@/entities/uniqueAssets';
 import { type UnlockableAppIconKey } from '@/features/app-icon/models/appIcons';
+import { type BackupFile, type CloudBackups } from '@/features/backup/backup';
 import { type ENSProfile } from '@/features/ens/types/profile';
 import { type REGISTRATION_MODES } from '@/features/ens/utils/helpers';
 import { type HlTrade, type PerpMarket, type PerpsPosition, type TriggerOrderSource, type TriggerOrderType } from '@/features/perps/types';
@@ -35,7 +36,6 @@ import type walletBackupStepTypes from '@/helpers/walletBackupStepTypes';
 import type WalletBackupTypes from '@/helpers/walletBackupTypes';
 import { type WalletConnectApprovalSheetType } from '@/helpers/walletConnectApprovalSheetTypes';
 import { type ChartTime } from '@/hooks/charts/useChartInfo';
-import { type BackupFile, type CloudBackups } from '@/model/backup';
 import { type RainbowWallet } from '@/model/wallet';
 import { type Route } from '@/navigation/routesNames';
 import type Routes from '@/navigation/routesNames';
