+++ src/screens/AddWalletSheet.tsx
@@ -15,9 +15,9 @@ import { type AddWalletItem } from '@/components/add-wallet/AddWalletRow';
 import useExperimentalFlag, { HARDWARE_WALLETS } from '@/config/experimentalHooks';
 import { Box, globalColors, Inset } from '@/design-system';
 import { IS_DEV } from '@/env';
+import { executeFnIfCloudBackupAvailable } from '@/features/backup/backup';
 import * as i18n from '@/languages';
 import { logger, RainbowError } from '@/logger';
-import { executeFnIfCloudBackupAvailable } from '@/model/backup';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
 import { type ImportFlowContext, type RootStackParamList } from '@/navigation/types';
