+++ src/screens/SettingsSheet/components/Backups/BackUpMenuButton.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useMemo, useState } from 'react';
 
 import { FloatingEmojis } from '@/components/floating-emojis';
 import Spinner from '@/components/Spinner';
+import { CloudBackupState } from '@/features/backup/stores/backupsStore';
 import useDimensions from '@/hooks/useDimensions';
 import * as i18n from '@/languages';
-import { CloudBackupState } from '@/state/backups/backups';
 import { useTheme } from '@/theme/ThemeContext';
 
 import MenuItem from '../MenuItem';
