+++ src/screens/SettingsSheet/components/SettingsSection.tsx
@@ -22,12 +22,12 @@ import { AppVersionStamp } from '@/components/AppVersionStamp';
 import { XIcon } from '@/components/icons/svg/XIcon';
 import useExperimentalFlag, { LANGUAGE_SETTINGS, NOTIFICATIONS } from '@/config/experimentalHooks';
 import { Box } from '@/design-system';
+import { backupsStore } from '@/features/backup/stores/backupsStore';
 import { showActionSheetWithOptions } from '@/framework/ui/utils/actionsheet';
 import walletBackupTypes from '@/helpers/walletBackupTypes';
 import useAccountSettings from '@/hooks/useAccountSettings';
 import * as i18n from '@/languages';
 import { useRemoteConfig } from '@/model/remoteConfig';
-import { backupsStore } from '@/state/backups/backups';
 import { useIsReadOnlyWallet } from '@/state/wallets/walletsStore';
 import { ReviewPromptAction } from '@/storage/schema';
 import { Themes, useTheme } from '@/theme/ThemeContext';
