+++ src/state/sync/SessionEntryPromptSync.tsx
@@ -5,6 +5,7 @@ import { type EthereumAddress } from '@/entities/wallet';
 import { IS_TEST } from '@/env';
 import { unlockableAppIcons, type UnlockableAppIconKey } from '@/features/app-icon/models/appIcons';
 import { unlockableAppIconCheck } from '@/features/app-icon/utils/unlockableAppIconCheck';
+import { backupsStore, LoadingStates, oneWeekInMs } from '@/features/backup/stores/backupsStore';
 import WalletBackupStepTypes from '@/helpers/walletBackupStepTypes';
 import walletBackupTypes from '@/helpers/walletBackupTypes';
 import WalletTypes from '@/helpers/walletTypes';
@@ -12,7 +13,6 @@ import { useActiveRoute } from '@/hooks/useActiveRoute';
 import { type RainbowAccount } from '@/model/wallet';
 import Navigation, { type Route } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
-import { backupsStore, LoadingStates, oneWeekInMs } from '@/state/backups/backups';
 import { isSwipeRoute } from '@/state/navigation/navigationStore';
 import { getSelectedWallet, getWallets, useWalletsStore } from '@/state/wallets/walletsStore';
 import { ReviewPromptAction } from '@/storage/schema';
